HUM-93: moved schemas to hooks
