+++ packages/apps/human-app/frontend/src/modules/worker/hooks/use-router-state.ts
@@ -1,5 +1,10 @@
+import { z } from 'zod';
 import { useLocationState } from '@/modules/worker/hooks/use-location-state';
-import { routerStateSchema } from '@/modules/worker/types/email-verification.types';
+
+export const routerStateSchema = z.object({
+  email: z.string().email(),
+  resendOnMount: z.boolean().optional(),
+});
 
 export function useRouterState() {
   const { field: routerState } = useLocationState({
